requirements.txt
pandas
datetime
google.auth.transport.requests
google.oauth2.id_token
google.oauth2.service_account
qdm
json
time
numpy
requests
certifi
urllib3
streamlit as st